fix(colour): moved colours into folder
